requirements.txt
mysql-connector-python
flask
pandas
numpy 
scikit-learn
streamlit
matplotlib